@@ -24,105 +25,118 @@
   "author": "Gnosis",
   "license": "ISC",
   "jest": {
-    "transform": {
-      "^.+\\.js$": "babel-jest",
-      "^.+\\.(css|less)$": "<rootDir>/node_modules/jest-css-modules"
-    },
     "modulePaths": [
       "<rootDir>/src/"
-    ]
+    ],
+    "moduleNameMapper": {
+      "\\.(css|scss)$": "identity-obj-proxy"
+    }
+  },
+  "resolutions": {
+    "**/bootstrap-loader/resolve": "1.6.0"
   },
   "pre-commit": [
     "precommit"
   ],
   "dependencies": {
-    "@gnosis.pm/gnosisjs": "1.1.0",
-    "autobind-decorator": "^1.4.1",
-    "await-delay": "^1.0.0",
+    "@gnosis.pm/gnosis-core-contracts": "^1.0.3",
+    "@gnosis.pm/gnosisjs": "^2.0.0",
+    "@gnosis.pm/olympia-token": "^1.3.0",
+    "autobind-decorator": "^2.1.0",
+    "babel-loader": "^8.0.0-beta.2",
     "babel-polyfill": "^6.23.0",
     "blockies": "0.0.2",
     "classnames": "^2.2.5",
     "cross-env": "^5.0.5",
-    "d3": "^4.9.1",
+    "d3": "^5.0.0",
     "d3-scale-chromatic": "^1.1.1",
-    "decimal.js": "^7.2.3",
-    "lodash": "4.17.4",
-    "moment": "^2.19.2",
-    "moment-duration-format": "^1.3.0",
+    "decimal.js": "^10.0.0",
+    "history": "^4.7.2",
+    "immutable": "^3.8.2",
+    "jsontokens": "^0.7.7",
+    "lodash": "^4.17.5",
+    "moment": "^2.22.0",
+    "moment-duration-format": "^2.2.1",
     "normalizr": "^3.2.2",
     "prop-types": "^15.6.0",
-    "raven-js": "^3.21.0",
     "raven-for-redux": "^1.1.1",
+    "raven-js": "^3.23.3",
     "rc-tooltip": "^3.5.0",
-    "react": "15.6.1",
-    "react-autosize-textarea": "^0.4.7",
-    "react-datetime": "^2.10.1",
-    "react-dom": "15.6.1",
-    "react-modal": "^2.4.1",
-    "react-redux": "^4.4.6",
-    "react-router": "3.0.2",
-    "react-router-redux": "^4.0.8",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1",
+    "react-immutable-proptypes": "^2.1.0",
+    "react-markdown": "^3.3.0",
+    "react-modal": "^3.1.11",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.9",
     "react-select": "^1.0.0-rc.10",
-    "react-textarea-autosize": "^4.0.7",
-    "react-transition-group": "^2.2.0",
+    "react-table": "^6.8.0",
+    "react-transition-group": "^2.3.0",
+    "react-youtube": "^7.5.0",
     "recharts": "^1.0.0-alpha.1",
+    "recompose": "^0.26.0",
     "redux": "^3.6.0",
     "redux-actions": "^2.0.1",
     "redux-form": "^7.1.2",
     "redux-thunk": "^2.2.0",
-    "reselect": "^2.5.4",
+    "reselect": "^3.0.1",
     "seedrandom": "^2.4.3",
     "sha1": "^1.1.1",
+    "uport-connect": "^0.7.1",
     "uuid": "^3.1.0",
-    "web3": "0.18.4"
+    "web3": "1.0.0-beta.30"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^3.2.12",
-    "@storybook/addon-knobs": "^3.2.12",
-    "@storybook/addon-links": "^3.2.12",
-    "@storybook/react": "^3.2.12",
-    "autoprefixer": "^7.1.6",
-    "babel-core": "6.26.0",
-    "babel-eslint": "^8.0.2",
-    "babel-loader": "7.1.2",
+    "@babel/cli": "^7.0.0-beta.39",
+    "@babel/core": "^7.0.0-beta.44",
+    "@babel/preset-env": "^7.0.0-beta.44",
+    "@babel/preset-react": "^7.0.0-beta.39",
+    "@babel/preset-stage-0": "^7.0.0-beta.39",
+    "@storybook/addon-actions": "^3.4.0",
+    "@storybook/addon-knobs": "^3.4.0",
+    "@storybook/addon-links": "^3.4.0",
+    "@storybook/react": "^3.4.0",
+    "autoprefixer": "^8.0.0",
+    "babel-core": "^7.0.0-0",
+    "babel-eslint": "^8.2.1",
+    "babel-jest": "^22.4.3",
     "babel-plugin-dynamic-import-node": "^1.2.0",
     "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-optional-chaining": "^7.0.0-beta.3",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-0": "6.24.1",
     "bootstrap-loader": "^2.1.0",
     "bootstrap-sass": "^3.3.7",
-    "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^0.27.3",
-    "eslint": "^4.11.0",
+    "case-sensitive-paths-webpack-plugin": "^2.1.1",
+    "css-loader": "^0.28.11",
+    "eslint": "^4.19.0",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-loader": "^1.9.0",
+    "eslint-loader": "^2.0.0",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jest": "^21.3.2",
+    "eslint-plugin-jest": "^21.15.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-react": "^7.4.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "^0.10.1",
-    "html-webpack-plugin": "^2.28.0",
-    "jest": "^20.0.4",
-    "jest-css-modules": "^1.1.0",
-    "json-loader": "0.5.4",
-    "less": "^2.7.2",
-    "less-loader": "^3.0.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "favicons-webpack-plugin": "0.0.9",
+    "file-loader": "^1.1.6",
+    "html-webpack-plugin": "^3.2.0",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^22.2.0",
+    "json-loader": "^0.5.7",
     "node-sass": "^4.5.3",
-    "postcss-loader": "^2.0.6",
+    "postcss-loader": "^2.1.2",
     "pre-commit": "^1.2.2",
-    "react-hot-loader": "^3.0.0-beta.7",
+    "raw-loader": "^0.5.1",
+    "react-hot-loader": "^4.0.1",
     "resolve-url-loader": "^2.0.3",
     "sass-loader": "^6.0.6",
-    "style-loader": "^0.13.2",
+    "style-loader": "^0.20.1",
     "truffle-hdwallet-provider": "0.0.3",
-    "uglifyjs-webpack-plugin": "^1.1.4",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.4"
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.5.0",
+    "webpack-cli": "^2.0.14",
+    "webpack-dev-server": "^3.1.1"
   }
 }
