@@ -1,33 +1,26 @@
 import {
-  getGnosisConnection,
   initGnosisConnection,
   getCurrentBalance,
   getCurrentAccount,
-  calcMarketGasCost,
-  calcBuySharesGasCost,
-  calcSellSharesGasCost,
-  calcCategoricalEventGasCost,
-  calcScalarEventGasCost,
-  calcCentralizedOracleGasCost,
-  calcFundingGasCost,
-  calcRedeemWinningsGasCost,
   getGasPrice,
-  getEtherTokens,
+  getTokenSymbol,
+  getTokenBalance,
+  initReadOnlyGnosisConnection,
 } from 'api'
+import Web3 from 'web3'
 
-import { timeoutCondition, getGnosisJsOptions, weiToEth } from 'utils/helpers'
-import { GAS_COST } from 'utils/constants'
+import { timeoutCondition, getGnosisJsOptions } from 'utils/helpers'
+import { findDefaultProvider } from 'integrations/store/selectors'
 import { createAction } from 'redux-actions'
-import { findDefaultProvider, getSelectedProvider } from 'selectors/blockchain'
-
-import { setActiveProvider } from './providers'
+import { setActiveProvider } from 'integrations/store/actions'
+import { getTokenAddress } from 'utils/configuration'
 
 // TODO define reducer for GnosisStatus
 export const setGnosisInitialized = createAction('SET_GNOSIS_CONNECTION')
 export const setConnectionStatus = createAction('SET_CONNECTION_STATUS')
-export const setGasCost = createAction('SET_GAS_COST')
 export const setGasPrice = createAction('SET_GAS_PRICE')
-export const setEtherTokens = createAction('SET_ETHER_TOKENS')
+export const setTokenBalance = createAction('SET_TOKEN_BALANCE')
+export const setTokenSymbol = createAction('SET_TOKEN_NAME')
 
 export const NETWORK_TIMEOUT = process.env.NODE_ENV === 'production' ? 10000 : 2000
 
@@ -36,45 +29,20 @@ export const requestGasPrice = () => async (dispatch) => {
   dispatch(setGasPrice({ entityType: 'gasPrice', gasPrice }))
 }
 
-export const requestGasCost = (contractType, opts) => async (dispatch) => {
-  if (contractType === GAS_COST.MARKET_CREATION) {
-    calcMarketGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.BUY_SHARES) {
-    calcBuySharesGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.SELL_SHARES) {
-    calcSellSharesGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.CATEGORICAL_EVENT) {
-    calcCategoricalEventGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.SCALAR_EVENT) {
-    calcScalarEventGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.CENTRALIZED_ORACLE) {
-    calcCentralizedOracleGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.FUNDING) {
-    calcFundingGasCost().then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
-  } else if (contractType === GAS_COST.REDEEM_WINNINGS) {
-    calcRedeemWinningsGasCost(opts).then((gasCost) => {
-      dispatch(setGasCost({ entityType: 'gasCosts', contractType, gasCost }))
-    })
+export const requestTokenSymbol = tokenAddress => async (dispatch) => {
+  let tokenSymbol
+  try {
+    tokenSymbol = await getTokenSymbol(tokenAddress)
+  } catch (e) {
+    tokenSymbol = ''
+  } finally {
+    dispatch(setTokenSymbol({ tokenAddress, tokenSymbol }))
   }
 }
 
-export const requestEtherTokens = account => async (dispatch) => {
-  const etherTokens = await getEtherTokens(account)
-  dispatch(setEtherTokens({ entityType: 'etherTokens', account, etherTokens }))
+export const requestTokenBalance = (tokenAddress, accountAddress) => async (dispatch) => {
+  const tokenBalance = await getTokenBalance(tokenAddress, accountAddress)
+  dispatch(setTokenBalance({ tokenAddress, tokenBalance }))
 }
 
 /**
@@ -82,42 +50,58 @@ export const requestEtherTokens = account => async (dispatch) => {
  */
 export const initGnosis = () => async (dispatch, getState) => {
   // initialize
+  let newProvider
+
   try {
     const state = getState()
 
     // determine new provider
-    const newProvider = findDefaultProvider(state)
+    newProvider = findDefaultProvider(state)
 
     if (newProvider) {
       await dispatch(setActiveProvider(newProvider.name))
       // init Gnosis connection
-      if (newProvider.account) {
-        const opts = getGnosisJsOptions(newProvider)
-        await initGnosisConnection(opts)
-      } else {
-        throw new Error('No account found')
-      }
 
+      const opts = getGnosisJsOptions(newProvider)
+      await initGnosisConnection(opts)
       await dispatch(setGnosisInitialized({ initialized: true }))
-      await requestEtherTokens()
+
+      if (newProvider.account) {
+        await getTokenBalance(getTokenAddress(), await getCurrentAccount())
+      }
     }
   } catch (error) {
     console.warn(`Gnosis.js initialization Error: ${error}`)
-    await dispatch(setConnectionStatus({ connected: false }))
+    dispatch(setConnectionStatus({ connected: false }))
     return dispatch(setGnosisInitialized({ initialized: false, error }))
   }
 
-  // connect
-  try {
-    // runs test executions on gnosisjs
-    const getConnection = async () => {
-      const account = await getCurrentAccount()
-      await getCurrentBalance(account)
+  if (newProvider) {
+    // connect
+    try {
+      // runs test executions on gnosisjs
+      const getConnection = async () => {
+        // these throw if they're not available, meaning we don't have a connection
+        const account = await getCurrentAccount()
+        await getCurrentBalance(account)
+      }
+
+      await Promise.race([getConnection(), timeoutCondition(NETWORK_TIMEOUT, 'connection timed out')])
+      await dispatch(setConnectionStatus({ connected: true }))
+    } catch (error) {
+      console.warn(`Gnosis.js connection Error: ${error}`)
+      return dispatch(setConnectionStatus({ connected: false }))
     }
-    await Promise.race([getConnection(), timeoutCondition(NETWORK_TIMEOUT, 'connection timed out')])
-    await dispatch(setConnectionStatus({ connected: true }))
+  }
+}
+
+export const initReadOnlyGnosis = () => async () => {
+  // initialize
+  try {
+    await initReadOnlyGnosisConnection({
+      ethereum: new Web3(new Web3.providers.HttpProvider(process.env.ETHEREUM_URL)).currentProvider,
+    })
   } catch (error) {
-    console.warn(`Gnosis.js connection Error: ${error}`)
-    return dispatch(setConnectionStatus({ connected: false }))
+    console.error(`Gnosis.js RO initialization Error: ${error}`)
   }
 }
