@@ -1,22 +1,35 @@
-import * as React from 'react'
+import React from 'react'
+import Markdown from 'react-markdown'
+import cn from 'classnames/bind'
 import Block from 'components/layout/Block'
 import Paragraph from 'components/layout/Paragraph'
-import './Footer.less'
+import { getFooterContent } from 'utils/configuration'
+import footerText from './footerText.txt'
+import style from './Footer.mod.scss'
 
-const Footer = () => (
-  <Block className="footer">
-    <Paragraph color="small">
-      Trading on prediction markets carry a high degree of financial risk to your capital. Please read our full{' '}
-      <a href="/RiskDisclaimerPolicy.html" target="_blank" className="footerLink">
-        Risk Disclaimer
-      </a>{' '}
-      and{' '}
-      <a href="/TermsOfService.html" target="_blank" className="footerLink">
-        Terms of Service
-      </a>{' '}
-      before trading
-    </Paragraph>
-  </Block>
-)
+const cx = cn.bind(style)
+const { type, source, markdown } = getFooterContent()
+
+const Footer = () => {
+  let text
+
+  if (type === 'text') {
+    text = source
+  } else if (type === 'file') {
+    text = footerText
+  }
+
+  return (
+    <Block margin="md">
+      {markdown ? (
+        <Markdown source={text} className={cx('footerContainer')} />
+      ) : (
+        <Paragraph center color="medium">
+          {source}
+        </Paragraph>
+      )}
+    </Block>
+  )
+}
 
 export default Footer
