@@ -1,93 +1,164 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'react-router'
+import { NavLink } from 'react-router-dom'
 import autobind from 'autobind-decorator'
-import { collateralTokenToText } from 'components/CurrencyName'
-import DecimalValue from 'components/DecimalValue'
-import Identicon from 'components/Identicon'
-import ProviderIcon from 'components/ProviderIcon'
-import { metamaskUnlocked } from 'integrations/utils'
-import { providerPropType } from 'utils/shapes'
 import { upperFirst } from 'lodash'
+import className from 'classnames/bind'
+import CurrencyName from 'components/CurrencyName'
+import DecimalValue from 'components/DecimalValue'
+import { providerPropType } from 'utils/shapes'
+import { shouldUseMetamask, shouldUseUport, areBadgesEnabled } from 'utils/configuration'
+import { hasMetamask } from 'integrations/metamask/utils'
 
+import Identicon from './Identicon'
+import ProviderIcon from './ProviderIcon'
+import BadgeIcon from './BadgeIcon'
 import MenuAccountDropdown from './MenuAccountDropdown'
 
-import './header.less'
+import css from './Header.scss'
+
+const cx = className.bind(css)
+
+const [useMetamask, useUport] = [shouldUseMetamask(), shouldUseUport()]
 
 class Header extends Component {
+  componentDidMount() {
+    if (this.props.isTournament && this.props.currentAccount) {
+      this.props.requestMainnetAddress()
+    }
+
+    if (this.props.currentAccount) {
+      this.props.requestTokenBalance(this.props.currentAccount)
+    }
+  }
+
+  componentDidUpdate(prevProps) {
+    // If user unlocks metamask, changes his account, we need to check if the account was registered
+    const shouldRequestMainnetAddress =
+      this.props.isTournament && this.props.currentAccount !== prevProps.currentAccount
+    if (shouldRequestMainnetAddress) {
+      this.props.requestMainnetAddress()
+    }
+  }
+
   @autobind
   async handleConnectWalletClick() {
-    if (!(await metamaskUnlocked())) {
-      this.props.openInstallMetamaskModal()
-    } else {
-      this.props.openAcceptTOSModal()
+    const { isConnectedToCorrectNetwork, lockedMetamask } = this.props
+    if (!hasMetamask() && !useUport) {
+      this.props.openModal('ModalInstallMetamask')
+    } else if (useMetamask) {
+      if (lockedMetamask) {
+        this.props.openModal('ModalUnlockMetamask')
+      } else if (!isConnectedToCorrectNetwork) {
+        this.props.openModal('ModalSwitchNetwork')
+      } else {
+        this.props.openModal('ModalRegisterWallet')
+      }
+    } else if (useUport) {
+      this.props.initUport()
     }
   }
 
   render() {
     const {
-      version, hasWallet, currentAccount, currentNetwork, currentBalance, currentProvider,
+      version,
+      hasWallet,
+      currentAccount,
+      currentNetwork,
+      tokenBalance,
+      currentProvider,
+      isTournament,
+      logoPath,
+      smallLogoPath,
+      showScoreboard,
+      showGameGuide,
+      gameGuideType,
+      gameGuideURL,
+      tokenAddress,
+      mainnetAddress,
+      userTournamentInfo,
     } = this.props
+
+    let walletConnected = hasWallet
+    if (isTournament && useMetamask) {
+      walletConnected = hasWallet && mainnetAddress
+    }
+
+    const logoVars = {}
+    if (isTournament) {
+      logoVars['--logoAnnotation'] = "'Powered by Gnosis'"
+      logoVars['--logoPath'] = `url("${logoPath}")`
+      logoVars['--smallLogoPath'] = `url("${smallLogoPath}")`
+    }
+
+    let gameGuideLink = <div />
+    if (showGameGuide) {
+      if (gameGuideType === 'default') {
+        gameGuideLink = (
+          <NavLink to="/game-guide" activeClassName={cx('active')} className={cx('navLink')}>
+            Game guide
+          </NavLink>
+        )
+      }
+
+      if (gameGuideType === 'link') {
+        gameGuideLink = (
+          <a href={gameGuideURL} className={cx('navLink')} target="_blank">
+            Game Guide
+          </a>
+        )
+      }
+    }
+
     return (
-      <div className="headerContainer">
-        <div className="container">
-          <div className="headerContainer__group headerContainer__group--logo">
-            <Link to="/">
-              <div className="headerLogo beta" />
-            </Link>
+      <div className={cx('headerContainer')}>
+        <div className={cx('container')}>
+          <div className={cx('group', 'logo')}>
+            <NavLink to="/markets/list">
+              <div className={cx('headerLogo', 'beta')} style={logoVars} />
+            </NavLink>
           </div>
-          <div className="headerContainer__group headerContainer__group--left headerContainer__group--version">
-            {version}
-          </div>
-          <div className="headerContainer__group headerContainer__group--left navLinks">
-            {hasWallet && (
-              <Link
-                to="/dashboard"
-                activeClassName="headerContainer__navLink--active"
-                className="headerContainer__navLink"
-              >
+          <div className={cx('group', 'left', 'version')}>{version}</div>
+          <div className={cx('group', 'left', 'navLinks')}>
+            {walletConnected && (
+              <NavLink to="/dashboard" activeClassName={cx('active')} className={cx('navLink')}>
                 Dashboard
-              </Link>
+              </NavLink>
             )}
-            <Link
-              to="/markets/list"
-              activeClassName="headerContainer__navLink--active"
-              className="headerContainer__navLink"
-            >
+            <NavLink to="/markets/list" activeClassName={cx('active')} className={cx('navLink')}>
               Markets
-            </Link>
-            {hasWallet && (
-              <Link
-                to="/transactions"
-                activeClassName="headerContainer__navLink--active"
-                className="headerContainer__navLink"
-              >
+            </NavLink>
+            {walletConnected && (
+              <NavLink to="/transactions" activeClassName={cx('active')} className={cx('navLink')}>
                 Transactions
-              </Link>
+              </NavLink>
             )}
+            {showScoreboard && (
+              <NavLink to="/scoreboard" activeClassName={cx('active')} className={cx('navLink')}>
+                Scoreboard
+              </NavLink>
+            )}
+            {gameGuideLink}
           </div>
 
-          <div className="headerContainer__group headerContainer__group--right">
-            {hasWallet &&
-              currentProvider && (
-              <div className="headerContainer__account">
-                {currentNetwork &&
-                    currentNetwork !== 'MAIN' && (
-                    <span className="headerContainer__network--text">
-                        Network: {upperFirst(currentNetwork.toLowerCase())}
-                  </span>
+          <div className={cx('group', 'right')}>
+            {walletConnected && currentProvider && (
+              <div className={cx('account')}>
+                {currentNetwork && currentNetwork !== 'MAIN' && (
+                  <span className={cx('network', 'text')}>Network: {upperFirst(currentNetwork.toLowerCase())}</span>
                 )}
-                <DecimalValue value={currentBalance} className="headerContainer__account--text" />&nbsp;
-                <span className="headerContainer__account--text">{collateralTokenToText()}</span>
+                <DecimalValue value={tokenBalance} className={cx('balance', 'test')} />&nbsp;
+                {tokenAddress ? <CurrencyName className={cx('account', 'text')} tokenAddress={tokenAddress} /> : <span>ETH</span>}
+                {areBadgesEnabled() && <BadgeIcon userTournamentInfo={userTournamentInfo} />}
                 <ProviderIcon provider={currentProvider} />
                 <Identicon account={currentAccount} />
-                <MenuAccountDropdown />
+                {useUport && <MenuAccountDropdown />}
               </div>
             )}
-            {!hasWallet && (
-              <a className="headerContainer__connect-wallet" onClick={this.handleConnectWalletClick}>
+            {!walletConnected && (
+              <button className={cx('connect-wallet')} onClick={this.handleConnectWalletClick}>
                 Connect a wallet
-              </a>
+              </button>
             )}
           </div>
         </div>
@@ -100,20 +171,41 @@ Header.propTypes = {
   version: PropTypes.string,
   currentNetwork: PropTypes.string,
   hasWallet: PropTypes.bool,
-  currentBalance: PropTypes.string,
+  tokenBalance: PropTypes.string,
   currentProvider: providerPropType,
   currentAccount: PropTypes.string,
-  openAcceptTOSModal: PropTypes.func.isRequired,
-  openInstallMetamaskModal: PropTypes.func.isRequired,
+  isTournament: PropTypes.bool,
+  userTournamentInfo: PropTypes.shape({}),
+  logoPath: PropTypes.string.isRequired,
+  smallLogoPath: PropTypes.string.isRequired,
+  showScoreboard: PropTypes.bool,
+  showGameGuide: PropTypes.bool,
+  gameGuideType: PropTypes.string,
+  gameGuideURL: PropTypes.string,
+  tokenAddress: PropTypes.string.isRequired,
+  lockedMetamask: PropTypes.bool,
+  requestMainnetAddress: PropTypes.func.isRequired,
+  requestTokenBalance: PropTypes.func.isRequired,
+  mainnetAddress: PropTypes.string,
+  initUport: PropTypes.func.isRequired,
+  openModal: PropTypes.func.isRequired,
 }
 
 Header.defaultProps = {
   version: '',
   currentNetwork: '',
   hasWallet: false,
-  currentBalance: '0',
+  tokenBalance: '0',
   currentProvider: {},
   currentAccount: '',
+  isTournament: false,
+  showScoreboard: false,
+  showGameGuide: false,
+  gameGuideType: 'default',
+  gameGuideURL: '',
+  mainnetAddress: undefined,
+  lockedMetamask: true,
+  userTournamentInfo: undefined,
 }
 
 export default Header
