@@ -1,10 +1,13 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 
-import './Notifications.less'
+import './Notifications.scss'
 
 const Notifications = ({ notifications, onClick }) => (
   <div className="notifications">
-    {notifications.map(({ id, title, icon, message }) => (
+    {notifications.map(({
+      id, title, icon, message,
+    }) => (
       <div key={id} className="notification" onClick={() => onClick()}>
         <div className="notification__title">
           {title}
@@ -16,9 +19,9 @@ const Notifications = ({ notifications, onClick }) => (
           {message}
         </div>
       </div>
-      ))}
+    ))}
   </div>
-  )
+)
 
 Notifications.propTypes = {
   notifications: PropTypes.arrayOf(PropTypes.object),
