@@ -6,7 +6,7 @@ import Span from 'components/layout/Span'
 
 import { getModerators, generateWalletName } from 'utils/helpers'
 
-import * as css from './style.less'
+import * as css from './style.mod.scss'
 
 const cx = classNames.bind(css)
 
@@ -44,13 +44,15 @@ class WalletAddress extends Component {
 
     const changeToVarText = showAddress ? 'deterministic name' : 'address'
     const text = this.getText()
-    return (<Span
-      className={cx('walletAddress')}
-      onClick={this.onAddressClick}
-      title={`${text} (Click to see ${changeToVarText})`}
-    >
-      { text }
-    </Span>)
+    return (
+      <Span
+        className={cx('walletAddress')}
+        onClick={this.onAddressClick}
+        title={`${text} (Click to see ${changeToVarText})`}
+      >
+        {text}
+      </Span>
+    )
   }
 }
 
