@@ -1,24 +1,27 @@
 /* global __VERSION__ */
 
 import React, { Component } from 'react'
+import { withRouter } from 'react-router-dom'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import TransitionGroup from 'react-transition-group/TransitionGroup'
 import CSSTransition from 'react-transition-group/CSSTransition'
-import { providerPropType } from 'utils/shapes'
-import Footer from 'components/Footer'
 
 import { connectBlockchain } from 'actions/blockchain'
-
-import LoadingIndicator from 'components/LoadingIndicator'
-
+import IndefiniteSpinner from 'components/Spinner/Indefinite'
 import PageFrame from 'components/layout/PageFrame'
-import TransactionFloaterContainer from 'containers/TransactionFloaterContainer'
+import Footer from 'components/Footer'
+import { providerPropType } from 'utils/shapes'
+import { hexWithoutPrefix } from 'utils/helpers'
 import HeaderContainer from 'containers/HeaderContainer'
+import TransactionFloaterContainer from 'containers/TransactionFloaterContainer'
+import { triedToConnect } from 'selectors/blockchain'
+import { getActiveProvider, isConnectedToCorrectNetwork } from 'integrations/store/selectors'
+import { shallDisplayFooter } from 'utils/configuration'
 
-import { getSelectedProvider, isConnectedToCorrectNetwork } from 'selectors/blockchain'
+import './app.scss'
 
-import './app.less'
+const displayFooter = shallDisplayFooter()
 
 class App extends Component {
   state = {
@@ -37,16 +40,16 @@ class App extends Component {
   }
 
   render() {
-    // if (!this.props.blockchainConnection) {
-    //   return (
-    //     <div className="appContainer">
-    //       <div className="loader-container">
-    //         <LoadingIndicator width={100} height={100} />
-    //         <h1>Connecting</h1>
-    //       </div>
-    //     </div>
-    //   )
-    // }
+    if (!this.props.blockchainConnection) {
+      return (
+        <div className="appContainer">
+          <div className="loader-container">
+            <IndefiniteSpinner width={100} height={100} />
+            <h1>Connecting</h1>
+          </div>
+        </div>
+      )
+    }
 
     const timeout = { enter: 200, exit: 200 }
 
@@ -70,9 +73,11 @@ class App extends Component {
         <HeaderContainer version={process.env.VERSION} />
         {this.props.provider && this.props.provider.account && <TransactionFloaterContainer />}
         {childrenContainer}
-        <PageFrame>
-          <Footer />
-        </PageFrame>
+        {displayFooter && (
+          <PageFrame>
+            <Footer />
+          </PageFrame>
+        )}
       </div>
     )
   }
@@ -85,12 +90,19 @@ App.propTypes = {
   provider: providerPropType,
 }
 
+App.defaultProps = {
+  blockchainConnection: false,
+  children: <div />,
+  location: {},
+  provider: {},
+}
+
 const mapStateToProps = state => ({
-  provider: getSelectedProvider(state),
-  blockchainConnection: state.blockchain.connectionTried,
+  provider: getActiveProvider(state),
+  blockchainConnection: triedToConnect(state),
   isConnectedToCorrectNetwork: isConnectedToCorrectNetwork(state),
 })
 
-export default connect(mapStateToProps, {
+export default withRouter(connect(mapStateToProps, {
   connectBlockchain,
-})(App)
+})(App))
