@@ -2,18 +2,19 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import cn from 'classnames'
 import { connect } from 'react-redux'
-
+import { withRouter } from 'react-router-dom'
+import { isTournament } from 'utils/configuration'
+import { triedToConnect } from 'selectors/blockchain'
 import { closeModal } from 'actions/modal'
+import * as modals from 'containers/Modals'
 
-import * as modals from 'containers/modals'
-
-import './backdrop.less'
+import './backdrop.scss'
 
 class BackdropProvider extends Component {
   renderBackdropContent() {
-    const { modal: { currentModal, isOpen, ...data }, closeModal: closeModalProp } = this.props
+    const { modal: { currentModal, isOpen, ...data }, closeModal: closeModalProp, blockchainConnection } = this.props
 
-    if (isOpen) {
+    if (isOpen && blockchainConnection) {
       /*
        * Implement more modals here by adding to components/modals.js
        */
@@ -30,18 +31,20 @@ class BackdropProvider extends Component {
   }
 
   render() {
-    const { children, modal: { isOpen } } = this.props
+    const { children, modal: { isOpen }, blockchainConnection } = this.props
+    const tournament = isTournament()
     return (
       <div className="backdrop">
         <div
           className={cn({
             backdrop__filter: true,
-            'backdrop__filter--visible': isOpen,
+            'backdrop__filter--visible': isOpen && blockchainConnection,
+            tournament,
           })}
         >
-          {isOpen ? <div style={{ position: 'fixed', minWidth: '100vw' }}>{children}</div> : children}
+          {isOpen && blockchainConnection ? <div style={{ position: 'fixed', minWidth: '100vw' }}>{children}</div> : children}
         </div>
-        <div className="backdrop__above">{this.renderBackdropContent()}</div>
+        <div className={cn('backdrop__above', { tournament })}>{this.renderBackdropContent()}</div>
       </div>
     )
   }
@@ -50,16 +53,18 @@ class BackdropProvider extends Component {
 BackdropProvider.propTypes = {
   modal: PropTypes.shape({
     isOpen: PropTypes.bool,
-  }),
-  children: PropTypes.node,
-  closeModal: PropTypes.func,
+  }).isRequired,
+  children: PropTypes.node.isRequired,
+  closeModal: PropTypes.func.isRequired,
+  blockchainConnection: PropTypes.bool.isRequired,
 }
 
 const mapStateToProps = state => ({
   modal: state.modal,
+  blockchainConnection: triedToConnect(state),
 })
 
 const mapDispatchToProps = dispatch => ({
   closeModal: () => dispatch(closeModal()),
 })
-export default connect(mapStateToProps, mapDispatchToProps)(BackdropProvider)
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(BackdropProvider))
