@@ -1,30 +1,9 @@
 import { connect } from 'react-redux'
-import { openModal } from 'actions/modal'
+import { withRouter } from 'react-router-dom'
 
 import Header from 'components/Header'
 
-import {
-  getCurrentBalance,
-  getCurrentNetwork,
-  getCurrentAccount,
-  getSelectedProvider,
-  checkWalletConnection,
-  isConnectedToCorrectNetwork,
-} from 'selectors/blockchain'
+import actions from './store/actions'
+import selectors from './store/selectors'
 
-const mapStateToProps = state => ({
-  hasWallet: checkWalletConnection(state),
-  currentAccount: getCurrentAccount(state),
-  currentBalance: getCurrentBalance(state),
-  currentNetwork: getCurrentNetwork(state),
-  currentProvider: getSelectedProvider(state),
-  isConnectedToCorrectNetwork: isConnectedToCorrectNetwork(state),
-})
-
-const mapDispatchToProps = dispatch => ({
-  openInstallMetamaskModal: () => dispatch(openModal({ modalName: 'ModalInstallMetamask' })),
-  openAcceptTOSModal: () => dispatch(openModal({ modalName: 'ModalAcceptTOS' })),
-  openNetworkCheckModal: () => dispatch(openModal({ modalName: 'ModalNetworkCheck' })),
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(Header)
+export default withRouter(connect(selectors, actions)(Header))
