@@ -18,10 +18,31 @@ class Metamask extends InjectedWeb3 {
 
   constructor() {
     super()
-    this.watcher = setInterval(() => {
+    this.watcher = () => {
       this.watch('account', this.getAccount)
       this.watch('balance', this.getBalance)
-    }, Metamask.watcherInterval)
+      this.watch('networkId', this.getNetworkId)
+    }
+  }
+
+  /**
+   * Tries to set connection to the blockchain
+   */
+  initWeb3() {
+    try {
+      if (
+        typeof window.web3 !== 'undefined' &&
+        (window.web3.currentProvider.constructor.name === 'MetamaskInpageProvider' ||
+          window.web3.currentProvider.isMetaMask)
+      ) {
+        this.web3 = new Web3(window.web3.currentProvider)
+        window.web3 = this.web3
+        return true
+      }
+      return false
+    } catch (err) {
+      return false
+    }
   }
 
   /**
@@ -34,18 +55,14 @@ class Metamask extends InjectedWeb3 {
     super.initialize(opts)
     this.runProviderRegister(this, { priority: Metamask.providerPriority })
 
-    this.walletEnabled = false
+    this.walletEnabled = this.initWeb3()
 
-    try {
-      if (hasMetamask()) {
-        this.web3 = new Web3(window.web3.currentProvider)
-        window.web3 = this.web3
-        this.walletEnabled = true
-      } else {
-        this.walletEnabled = false
-      }
-    } catch (err) {
-      this.walletEnabled = false
+    if (this.watcher) {
+      setInterval(this.watcher, Metamask.watcherInterval)
+    }
+
+    if (this.watcher) {
+      setInterval(this.watcher, Metamask.watcherInterval)
     }
 
     if (this.walletEnabled) {
