@@ -1,5 +1,7 @@
 import { get } from 'lodash'
-import { entitySelector } from './entities'
+import { entitySelector } from 'selectors/entities'
+import { getMarketById } from 'selectors/market'
+import { hexWithPrefix } from 'utils/helpers'
 
 export const EVENT_TYPE_CATEGORICAL = 'CATEGORICAL'
 export const EVENT_TYPE_SCALAR = 'SCALAR'
@@ -24,4 +26,44 @@ export const getEvents = (state) => {
   }
 
   return state.entities.events
+}
+
+export const eventMarketsSelector = (state) => {
+  if (!state.entities) {
+    return {}
+  }
+
+  if (!state.entities.markets) {
+    return {}
+  }
+
+  const { markets } = state.entities
+  const eventMarkets = {}
+
+  const marketSelector = getMarketById(state)
+
+  Object.keys(markets).forEach((marketAddress) => {
+    eventMarkets[markets[marketAddress].event] = marketSelector(marketAddress)
+  })
+
+  return eventMarkets
+}
+
+export const eventMarketSelector = marketAddress => (state) => {
+  if (!state.entities) {
+    return {}
+  }
+
+  if (!state.entities.markets) {
+    return {}
+  }
+
+  if (!state.entities.markets[marketAddress]) {
+    return {}
+  }
+
+  const market = getMarketById(state)(marketAddress)
+  const eventAddress = hexWithPrefix(market.event.address)
+
+  return { [eventAddress]: market }
 }
