@@ -2,14 +2,15 @@ import { createSelector } from 'reselect'
 import Decimal from 'decimal.js'
 import moment from 'moment'
 
-import { add0xPrefix, calcShareWinnings } from 'utils/helpers'
+import { add0xPrefix } from 'utils/helpers'
+import { calcShareWinnings } from 'containers/DashboardPage/store/selectors/utils'
 import { calcLMSRMarginalPrice, calcLMSROutcomeTokenCount } from 'api'
 
-import { getCurrentAccount } from './blockchain'
-import { getEvents } from './event'
-import { getOracles } from './oracle'
-import { getEventDescriptions } from './eventDescription'
-import { MARKET_STAGES } from '../utils/constants'
+import { getCurrentAccount } from 'integrations/store/selectors'
+import { getEvents } from 'selectors/event'
+import { getOracles } from 'selectors/oracle'
+import { getEventDescriptions } from 'selectors/eventDescription'
+import { MARKET_STAGES } from 'utils/constants'
 
 export const getShares = (state) => {
   if (!state.entities) {
@@ -42,26 +43,7 @@ const eventMarketsSelector = (state) => {
   return eventMarkets
 }
 
-const eventMarketSelector = marketAddress => (state) => {
-  if (!state.entities) {
-    return {}
-  }
-
-  if (!state.entities.markets) {
-    return {}
-  }
-
-  if (!state.entities.markets[marketAddress]) {
-    return {}
-  }
-
-  const market = state.entities.markets[marketAddress]
-  const eventAddress = add0xPrefix(market.event)
-
-  return { [eventAddress]: market }
-}
-
-const eventSharesSelector = createSelector(getCurrentAccount, getShares, (account, shares) => {
+export const eventSharesSelector = createSelector(getCurrentAccount, getShares, (account, shares) => {
   const eventShares = {}
 
   Object.keys(shares).forEach((shareId) => {
@@ -83,7 +65,7 @@ const eventSharesSelector = createSelector(getCurrentAccount, getShares, (accoun
   return eventShares
 })
 
-const enhanceShares = (oracles, events, eventDescriptions, eventMarkets, eventShares, account) => {
+export const enhanceShares = (oracles, events, eventDescriptions, eventMarkets, eventShares) => {
   const enhancedShares = {}
 
   Object.keys(eventShares).forEach((eventAddress) => {
@@ -106,7 +88,7 @@ const enhanceShares = (oracles, events, eventDescriptions, eventMarkets, eventSh
       const isShareEventResolved = oracle.isOutcomeSet && event.isWinningOutcomeSet
       const isShareMarketClosed =
         market.stage === MARKET_STAGES.MARKET_CLOSED || moment(resolutionDate).isBefore(moment().utc())
-      const shareWinning = isShareEventResolved ? calcShareWinnings(share, market, event, account) : '0'
+      const shareWinning = isShareEventResolved ? calcShareWinnings(share, market, event) : '0'
 
       if (isShareEventResolved && Decimal(shareWinning).eq(0)) {
         return
@@ -168,13 +150,15 @@ export const getAccountShares = createSelector(
   enhanceShares,
 )
 
-export const getMarketShares = marketAddress =>
-  createSelector(
-    getOracles,
-    getEvents,
-    getEventDescriptions,
-    eventMarketSelector(marketAddress),
-    eventSharesSelector,
-    getCurrentAccount,
-    enhanceShares,
-  )
+export const getRedeemedShares = (state, marketAddress) => {
+  const shares = getAccountShares(state)
+
+  const redeemedShares = {}
+  Object.keys(shares).forEach((shareId) => {
+    const share = shares[shareId]
+    if (share.market && share.market.address === marketAddress) {
+      redeemedShares[shareId] = share
+    }
+  })
+  return redeemedShares
+}
