@@ -5,16 +5,11 @@ import moment from 'moment'
 import { HEX_VALUE_REGEX, OUTCOME_TYPES, MARKET_STAGES } from 'utils/constants'
 import { WALLET_PROVIDER } from 'integrations/constants'
 import Web3 from 'web3'
+import Uport from 'integrations/uport'
 
 import dictionary from 'randomNames.json'
 
-export const hexWithoutPrefix = (value) => {
-  if (HEX_VALUE_REGEX.test(value)) {
-    return startsWith(value, '0x') ? value.substring(2) : value
-  }
-
-  return value
-}
+export const hexWithoutPrefix = value => (startsWith(value, '0x') ? value.substring(2) : value)
 
 /**
  * Adds the `0x` prefix to the incoming string value
@@ -47,7 +42,7 @@ export const toEntity = (data, entityType, idKey = 'address') => {
 
 /**
  * Converts a value from WEI to ETH
- * @param {String|Number} value
+ * @param {String|Number|Decimal} value
  */
 export const weiToEth = (value) => {
   let ethValue = new Decimal(0)
@@ -94,17 +89,6 @@ export const normalizeScalarPoint = (
     .add(bigLowerBound)
     .toDP(decimals)
     .toNumber()
-}
-
-/**
- * Adds _id incremental numeric property to each object in the array
- * @param { objects[] } arrayData
- */
-export const addIdToObjectsInArray = (arrayData) => {
-  arrayData.forEach((item, index) => {
-    item._id = index
-  })
-  return arrayData
 }
 
 export const restFetch = url =>
@@ -164,6 +148,10 @@ export const getGnosisJsOptions = (provider) => {
   } else if (provider && provider === WALLET_PROVIDER.PARITY) {
     // Inject window.web3
     opts.ethereum = window.web3.currentProvider
+  } else if (provider && provider.name === WALLET_PROVIDER.UPORT) {
+    const { uport } = Uport
+    opts.ethereum = uport.getProvider()
+    opts.defaultAccount = provider.account
   } else {
     // Default remote node
     opts.ethereum = new Web3(new Web3.providers.HttpProvider(`${process.env.ETHEREUM_URL}`)).currentProvider
@@ -173,20 +161,6 @@ export const getGnosisJsOptions = (provider) => {
   return opts
 }
 
-export const promisify = (func, params, timeout) =>
-  new Promise((resolve, reject) => {
-    if (timeout) {
-      setTimeout(() => reject(new Error('Promise timed out')), timeout)
-    }
-
-    func(...params, (err, res) => {
-      if (err) {
-        return reject(err)
-      }
-      return resolve(res)
-    })
-  })
-
 const BLOCKED_WORD_LIST = ['sexual', 'african', 'american', 'european', 'asian', 'israeli']
 
 export const generateDeterministicRandomName = (seed) => {
@@ -194,7 +168,7 @@ export const generateDeterministicRandomName = (seed) => {
 
   // generate 5 so later we have more params left over for longer names, without changing everyone's other words
   // eslint-disable-next-line no-unsued-vars
-  const [r1, r2, ...params] = range(0, 5).map(rng)
+  const [r1, r2] = range(0, 5).map(rng)
 
   const { adjectives, nouns } = dictionary
 
@@ -221,95 +195,4 @@ export const generateWalletName = (account) => {
   const accountAddressNormalized = hexWithPrefix(account).toLowerCase()
 
   return generateDeterministicRandomName(accountAddressNormalized)
-}
-
-const isValidMarket = market => !!(market && market.event && market.oracle && market.eventDescription)
-
-const marketCanRedeemWinnings = market => market.event.isWinningOutcomeSet
-
-const SCALAR_OUTCOME_RANGE = 1000000
-
-export const getMarketWinningsCategorical = (market, shares, account) => {
-  if (!isValidMarket(market) || !marketCanRedeemWinnings(market)) {
-    return {}
-  }
-
-  const marketOutcome = parseInt(market.event.outcome, 10)
-
-  // let winnings = Decimal(0)
-  const winningsByOutcome = {}
-
-  shares.forEach((share) => {
-    const shareOutcome = parseInt(share.outcomeToken.index, 10)
-    const belongsToMarket = share.outcomeToken.event === market.event.address
-    const belongsToUser = share.owner === account
-    const outcomeWon = shareOutcome === marketOutcome
-    if (belongsToMarket && belongsToUser && outcomeWon) {
-      const outcomeInt = parseInt(share.outcomeToken.event, 10)
-
-      // multiple shares bought for same outcome
-      if (!winningsByOutcome[outcomeInt]) {
-        winningsByOutcome[outcomeInt] = Decimal(0)
-      }
-      winningsByOutcome[outcomeInt] = winningsByOutcome[outcomeInt].add(Decimal(share.balance))
-    }
-  })
-
-  // object length will always be 1 if won
-  return mapValues(winningsByOutcome, val => val.toString())
-}
-
-export const calcShareWinningsCategorical = (share, market, event) => {
-  const outcome = parseInt(event.outcome, 10)
-  const shareOutcome = parseInt(share.outcomeToken.index, 10)
-  if (shareOutcome !== outcome) {
-    return '0'
-  }
-
-  return Decimal(share.balance).toString()
-}
-
-export const calcShareWinningsScalar = (share, market, event) => {
-  const outcomeRange = Decimal(SCALAR_OUTCOME_RANGE)
-  const outcome = Decimal(parseInt(event.outcome, 10))
-  const lowerBound = Decimal(event.lowerBound)
-  const upperBound = Decimal(event.upperBound)
-
-  let outcomeClamped = Decimal(0)
-
-  if (outcome.lt(lowerBound)) {
-    outcomeClamped = Decimal(0)
-  } else if (outcome.gt(upperBound)) {
-    outcomeClamped = outcomeRange
-  } else {
-    outcomeClamped = outcomeRange.mul(outcome.sub(lowerBound).toString()).div(upperBound.sub(lowerBound).toString())
-  }
-
-  const factorShort = outcomeRange.sub(outcomeClamped)
-  const factorLong = outcomeRange.sub(factorShort.toString())
-
-  const isShort = parseInt(share.outcomeToken.index, 10) === 0
-  const isLong = parseInt(share.outcomeToken.index, 10)
-  if (isShort) {
-    return Decimal(share.balance)
-      .mul(factorShort)
-      .div(outcomeRange)
-  }
-
-  if (isLong) {
-    return Decimal(share.balance)
-      .mul(factorLong)
-      .div(outcomeRange)
-      .toString()
-  }
-
-  throw new Error(`Invalid Outcome for Scalar Event found: ${share.outcomeToken.index}`)
-}
-
-export const calcShareWinnings = (share, market, event) => {
-  const isCategorical = event.type === OUTCOME_TYPES.CATEGORICAL
-
-  return isCategorical
-    ? calcShareWinningsCategorical(share, market, event)
-    : calcShareWinningsScalar(share, market, event)
 }
