@@ -8,6 +8,8 @@ export const eventDescriptionShape = PropTypes.shape({
   outcomes: PropTypes.arrayOf(PropTypes.string),
   resolutionDate: PropTypes.string,
   title: PropTypes.string,
+  decimals: PropTypes.number,
+  unit: PropTypes.string,
 })
 
 export const outcomeTokenShape = PropTypes.shape({
@@ -15,14 +17,6 @@ export const outcomeTokenShape = PropTypes.shape({
   event: PropTypes.string,
   index: PropTypes.number,
   totalSupply: PropTypes.string,
-})
-
-export const marketShape = PropTypes.shape({
-  event: PropTypes.object,
-  eventDescription: eventDescriptionShape,
-  oracle: PropTypes.object,
-  netOutcomeTokensSold: PropTypes.arrayOf(PropTypes.string),
-  address: PropTypes.string,
 })
 
 export const providerPropType = PropTypes.shape({
@@ -38,6 +32,17 @@ export const eventShape = PropTypes.shape({
   event: PropTypes.string,
   label: PropTypes.string,
   status: PropTypes.string,
+  type: PropTypes.string,
+  lowerBound: PropTypes.string,
+  upperBound: PropTypes.string,
+})
+
+export const marketShape = PropTypes.shape({
+  event: eventShape,
+  eventDescription: eventDescriptionShape,
+  oracle: PropTypes.object,
+  netOutcomeTokensSold: PropTypes.arrayOf(PropTypes.string),
+  address: PropTypes.string,
 })
 
 export const marketShareShape = PropTypes.shape({
@@ -66,15 +71,9 @@ export const transactionShape = PropTypes.shape({
   completionStatus: PropTypes.oneOf(Object.values(TRANSACTION_COMPLETE_STATUS)),
 })
 
-export const gasCostsShape = PropTypes.shape({
-  redeemWinnings: PropTypes.number,
-  market: PropTypes.number,
-  buyShares: PropTypes.number,
-  sellShares: PropTypes.number,
-  resolveOracle: PropTypes.number,
-  withdrawFees: PropTypes.number,
-  categoricalEvent: PropTypes.number,
-  scalarEvent: PropTypes.number,
-  centralizedOracle: PropTypes.number,
-  funding: PropTypes.number,
+export const ReactRouterMatchShape = PropTypes.shape({
+  isExact: PropTypes.bool,
+  path: PropTypes.string,
+  url: PropTypes.string,
+  params: PropTypes.object,
 })
