@@ -1,6 +1,7 @@
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const FaviconsWebpackPlugin = require('favicons-webpack-plugin')
-
+const CaseSensitivePathsPlugin = require('case-sensitive-paths-webpack-plugin')
+const config = require('./src/config.json')
 const path = require('path')
 const webpack = require('webpack')
 const pkg = require('./package.json')
@@ -8,20 +9,25 @@ const pkg = require('./package.json')
 const version = process.env.BUILD_VERSION || pkg.version
 const build = process.env.BUILD_NUMBER || 'SNAPSHOT'
 
-const config = require('./src/config.json')
-
+const isTournament = config.interface && config.interface.tournament
+const defaultFavicon = isTournament ? 'assets/img/gnosis_apollo_favicon.png' : 'assets/img/gnosis_logo_favicon.png'
+const faviconPath =
+  config.interface && config.interface.faviconPath && isTournament ? config.interface.faviconPath : defaultFavicon
 const whitelist = config.developmentWhitelist
 
 const gnosisDbUrl =
-  process.env.GNOSISDB_URL || `${config.gnosisdb.protocol}://${config.gnosisdb.hostDev}${config.gnosisdb.port ? `:${config.gnosisdb.port}` : ''}`
+  process.env.GNOSISDB_URL ||
+  `${config.gnosisdb.protocol}://${config.gnosisdb.hostDev}${config.gnosisdb.port ? `:${config.gnosisdb.port}` : ''}`
 
 const ethereumUrl =
-  process.env.ETHEREUM_URL || `${config.ethereum.protocol}://${config.ethereum.hostDev}${config.ethereum.port ? `:${config.ethereum.port}` : ''}`
+  process.env.ETHEREUM_URL ||
+  `${config.ethereum.protocol}://${config.ethereum.hostDev}${config.ethereum.port ? `:${config.ethereum.port}` : ''}`
 
 module.exports = {
   context: path.join(__dirname, 'src'),
-  entry: ['react-hot-loader/patch', 'bootstrap-loader', 'index.js'],
+  entry: ['bootstrap-loader', 'index.js'],
   devtool: 'eval-source-map',
+  mode: 'development',
   output: {
     publicPath: '/',
     path: `${__dirname}/dist`,
@@ -48,14 +54,20 @@ module.exports = {
         test: /\.(jpe?g|png|svg)$/i,
         loader: 'file-loader?hash=sha512&digest=hex&name=img/[hash].[ext]',
       },
+      // TODO: Remove this special rule for css-modules when all globally scoped CSS is removed
+      // change the RegEx to: `/*.(scss|css)$/`
       {
-        test: /\.(less|css)$/,
+        test: /\.mod\.(scss|css)$/,
         use: [
-          {
-            loader: 'style-loader',
-          },
+          'style-loader',
           {
             loader: 'css-loader',
+            options: {
+              sourceMap: true,
+              modules: true,
+              localIdentName: '[name]__[local]___[hash:base64:5]',
+              importLoaders: 2,
+            },
           },
           {
             loader: 'postcss-loader',
@@ -63,18 +75,46 @@ module.exports = {
               sourceMap: true,
             },
           },
-          { loader: 'less-loader', options: { strictMath: true } },
+          { loader: 'sass-loader', options: { sourceMap: true } },
+        ],
+      },
+      {
+        test: /^((?!\.mod).)*\.(css|scss)$/,
+        use: [
+          'style-loader',
+          {
+            loader: 'css-loader',
+            options: {
+              sourceMap: true,
+              importLoaders: 2,
+            },
+          },
+          {
+            loader: 'postcss-loader',
+            options: {
+              sourceMap: true,
+            },
+          },
+          {
+            loader: 'sass-loader',
+            options: { sourceMap: true, includePaths: [path.resolve(__dirname, './src')] },
+          },
         ],
       },
       {
         test: /\.(ttf|otf|eot|woff(2)?)(\?[a-z0-9]+)?$/,
         loader: 'file-loader?name=fonts/[name].[ext]',
       },
+      {
+        test: /\.txt$/,
+        use: 'raw-loader',
+      },
     ],
   },
   devServer: {
     disableHostCheck: true,
     historyApiFallback: true,
+    hot: true,
     port: 5000,
     proxy: {
       '/api': {
@@ -85,11 +125,12 @@ module.exports = {
     watchOptions: {
       ignored: /node_modules/,
     },
-    contentBase: [path.join(__dirname, 'assets'), path.join(__dirname, 'dist')],
+    contentBase: [path.join(__dirname, 'dist'), path.join(__dirname, 'src')],
   },
   plugins: [
+    new CaseSensitivePathsPlugin(),
     new FaviconsWebpackPlugin({
-      logo: 'assets/img/gnosis_logo_favicon.png',
+      logo: faviconPath,
       // Generate a cache file with control hashes and
       // don't rebuild the favicons until those hashes change
       persistentCache: true,
@@ -121,5 +162,6 @@ module.exports = {
       TRAVIS_BUILD_ID: null,
       TRAVIS_BRANCH: null,
     }),
+    new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /en/),
   ],
 }
